Moving beacon info structure into message package
